requirements.txt
Django==1.8.3
django-cors-headers==1.1.0
django-extensions==1.5.5
djangorestframework==3.2.0
psycopg2==2.6.1
requests==2.7.0
six==1.9.0
